Code reorganization

+++ moa/src/main/java/com/github/javacliparser/gui/IntOptionEditComponent.java
@@ -15,7 +15,7 @@
  *
  *    You should have received a copy of the GNU General Public License
  *    along with this program. If not, see <http://www.gnu.org/licenses/>.
- *    
+ *
  */
 package com.github.javacliparser.gui;
 
@@ -37,58 +37,57 @@ import com.github.javacliparser.Option;
  * @author Richard Kirkby (rkirkby@cs.waikato.ac.nz)
  * @version $Revision: 7 $
  */
-public class IntOptionEditComponent extends JPanel implements
-        OptionEditComponent {
+public class IntOptionEditComponent extends JPanel implements OptionEditComponent {
 
-    private static final long serialVersionUID = 1L;
+	private static final long serialVersionUID = 1L;
 
-    protected IntOption editedOption;
+	protected IntOption editedOption;
 
-    protected JSpinner spinner;
+	protected JSpinner spinner;
 
-    protected JSlider slider;
+	protected JSlider slider;
 
-    public IntOptionEditComponent(Option opt) {
-        IntOption option = (IntOption) opt;
-        this.editedOption = option;
-        int minVal = option.getMinValue();
-        int maxVal = option.getMaxValue();
-        setLayout(new GridLayout(1, 0));
-        this.spinner = new JSpinner(new SpinnerNumberModel(option.getValue(),
-                minVal, maxVal, 1));
-        add(this.spinner);
-        if ((minVal > Integer.MIN_VALUE) && (maxVal < Integer.MAX_VALUE)) {
-            this.slider = new JSlider(minVal, maxVal, option.getValue());
-            add(this.slider);
-            this.slider.addChangeListener(new ChangeListener() {
+	public IntOptionEditComponent(Option opt) {
+		IntOption option = (IntOption) opt;
+		this.editedOption = option;
+		int minVal = option.getMinValue();
+		int maxVal = option.getMaxValue();
+		setLayout(new GridLayout(1, 0));
+		this.spinner = new JSpinner(new SpinnerNumberModel(option.getValue(), minVal, maxVal, 1));
+		add(this.spinner);
+		if ((minVal > Integer.MIN_VALUE) && (maxVal < Integer.MAX_VALUE)) {
+			this.slider = new JSlider(minVal, maxVal, option.getValue());
+			add(this.slider);
+			this.slider.addChangeListener(new ChangeListener() {
 
-                @Override
-                public void stateChanged(ChangeEvent e) {
-                    IntOptionEditComponent.this.spinner.setValue(IntOptionEditComponent.this.slider.getValue());
-                }
-            });
-            this.spinner.addChangeListener(new ChangeListener() {
+				@Override
+				public void stateChanged(ChangeEvent e) {
+					IntOptionEditComponent.this.spinner.setValue(IntOptionEditComponent.this.slider.getValue());
+				}
+			});
+			this.spinner.addChangeListener(new ChangeListener() {
 
-                @Override
-                public void stateChanged(ChangeEvent e) {
-                    IntOptionEditComponent.this.slider.setValue(((Integer) IntOptionEditComponent.this.spinner.getValue()).intValue());
-                }
-            });
-        }
-    }
+				@Override
+				public void stateChanged(ChangeEvent e) {
+					IntOptionEditComponent.this.slider
+							.setValue(((Integer) IntOptionEditComponent.this.spinner.getValue()));
+				}
+			});
+		}
+	}
 
-    @Override
-    public void applyState() {
-        this.editedOption.setValue(((Integer) this.spinner.getValue()).intValue());
-    }
+	@Override
+	public void applyState() {
+		this.editedOption.setValue(((Integer) this.spinner.getValue()));
+	}
 
-    @Override
-    public Option getEditedOption() {
-        return this.editedOption;
-    }
+	@Override
+	public Option getEditedOption() {
+		return this.editedOption;
+	}
 
-    @Override
-    public void setEditState(String cliString) {
-        this.spinner.setValue(IntOption.cliStringToInt(cliString));
-    }
+	@Override
+	public void setEditState(String cliString) {
+		this.spinner.setValue(IntOption.cliStringToInt(cliString));
+	}
 }
